LayoutTests/fast/dom/Geolocation/permission-denied-expected.txt

Tests Geolocation when permission is denied, using the mock service.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User disallowed Geolocation"
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

TEST COMPLETE
